feat: remove openapi and metrcs middleware
